entity is now abstract class; logging moved to generic; method defaults
